pyproject.toml
[tool.poetry]
name = "mynebackend"
version = "0.1.0"
description = ""
authors = ["mhmad wrekat <mhmmadwreekkat@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.1.7"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
django-environ = "^0.8.1"
djangorestframework = "^3.13.1"
pytz = "^2021.3"
django-cors-headers = "^3.10.1"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
